bind/java: Do not get ApplicationContext via reflection, ask for it when needed.
Previously any library created using gobind used reflection to get the ApplicationContext. The methods used are
disallowed in Android 9 causing any library to crash.  However the ApplictionContext is not required and only needed
when using RunOnJvm. This change adds an method to set the ApplicationContext when needed, but will not get it by
default by using reflection.

Fixes #31364