bind: Fix setter name generation for ObjC
The generation for getters and setters of Go
exported fields that followed the acronyms
rule, i.e. for a field named `OSName` the
resulting getter and setter were `osName` and
`setOSName` respectively, however the Obj-C rules
defined here:

  * https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/ProgrammingWithObjectiveC/EncapsulatingData/EncapsulatingData.html

expected for the names to be the following: `osName`
and `setOsName`.

This commit updates setter name generation by capitalizing
the first char and lower casing the second one, it leaves
the getter name unchaged, this also updates the golden files.

Fix golang/go#32008